(refactor): adjust imports with alias paths
