+++ src/components/event-info/index.tsx
