+++ src/components/hero/index.ts
