+++ src/components/lectures/LectureCard.tsx
