+++ src/components/lectures/index.ts
