+++ src/components/primitives/carousel.tsx
@@ -3,8 +3,8 @@ import useEmblaCarousel, {
   type UseEmblaCarouselType,
 } from "embla-carousel-react";
 
-import { cn } from "../lib/utils";
-import { Button } from "./button";
+import { cn } from "@/lib/utils";
+import { Button } from "@/components/primitives/button";
 
 type CarouselApi = UseEmblaCarouselType[1];
 type UseCarouselParameters = Parameters<typeof useEmblaCarousel>;
