+++ src/components/sponsors/index.tsx
