+++ vite.config.ts
@@ -1,13 +1,27 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-import path from 'path'
+import { fileURLToPath, URL } from "url";
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+import path from "path";
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
   resolve: {
-    alias: {
-      '@/': path.resolve(__dirname, 'src'),
-    }
-  }
-})
+    alias: [
+      {
+        find: "@/components",
+        replacement: fileURLToPath(
+          new URL("./src/components", import.meta.url)
+        ),
+      },
+      {
+        find: "@/pages",
+        replacement: fileURLToPath(new URL("./src/pages", import.meta.url)),
+      },
+      {
+        find: "@/lib",
+        replacement: fileURLToPath(new URL("./src/lib", import.meta.url)),
+      },
+    ],
+  },
+});
